requirements.txt
# Core App Requirements
gradio==4.24.0
paddleocr==2.6.1.3
paddlepaddle==2.6.2
Pillow
numpy
requests

# LLM Client
ollama==0.1.7
